Add Swagger export as one API declaration (restlet/apispark#3876)
