+++ modules/org.restlet.ext.swagger/src/org/restlet/ext/swagger/SwaggerSpecificationRestlet.java
@@ -32,7 +32,7 @@ import org.restlet.Restlet;
 import org.restlet.data.Method;
 import org.restlet.data.Reference;
 import org.restlet.data.Status;
-import org.restlet.ext.apispark.internal.conversion.swagger.v1_2.SwaggerTranslator;
+import org.restlet.ext.apispark.internal.conversion.swagger.v1_2.SwaggerWriter;
 import org.restlet.ext.apispark.internal.conversion.swagger.v1_2.model.ApiDeclaration;
 import org.restlet.ext.apispark.internal.conversion.swagger.v1_2.model.ResourceListing;
 import org.restlet.ext.apispark.internal.introspection.application.ApplicationIntrospector;
@@ -91,9 +91,9 @@ public class SwaggerSpecificationRestlet extends Restlet {
 
     /**
      * The version of the Swagger specification. Default is
-     * {@link SwaggerTranslator#SWAGGER_VERSION}
+     * {@link SwaggerWriter#SWAGGER_VERSION}
      */
-    private String swaggerVersion = SwaggerTranslator.SWAGGER_VERSION;
+    private String swaggerVersion = SwaggerWriter.SWAGGER_VERSION;
 
     /**
      * Default constructor.<br>
@@ -165,8 +165,7 @@ public class SwaggerSpecificationRestlet extends Restlet {
      * @return The representation of the API declaration.
      */
     public Representation getApiDeclaration(String category) {
-        ApiDeclaration apiDeclaration = SwaggerTranslator.getApiDeclaration(
-                category, getDefinition());
+        ApiDeclaration apiDeclaration = SwaggerWriter.getApiDeclaration(getDefinition(), category);
         apiDeclaration.setSwaggerVersion(swaggerVersion);
         return new JacksonRepresentation<>(apiDeclaration);
     }
@@ -218,15 +217,14 @@ public class SwaggerSpecificationRestlet extends Restlet {
      *         Application.
      */
     public Representation getResourceListing() {
-        ResourceListing resourcelisting = SwaggerTranslator
-                .getResourcelisting(getDefinition());
+        ResourceListing resourcelisting = SwaggerWriter.getResourcelisting(getDefinition());
         resourcelisting.setSwaggerVersion(swaggerVersion);
         return new JacksonRepresentation<>(resourcelisting);
     }
 
     /**
      * Returns the version of the Swagger specification. Default is
-     * {@link SwaggerTranslator#SWAGGER_VERSION}
+     * {@link SwaggerWriter#SWAGGER_VERSION}
      * 
      * @return The version of the Swagger specification.
      */
